+++ src/cdc_fifo_gray.sv
@@ -1,4 +1,4 @@
-// Copyright 2018 ETH Zurich and University of Bologna.
+// Copyright 2018-2019 ETH Zurich and University of Bologna.
 //
 // Copyright and related rights are licensed under the Solderpad Hardware
 // License, Version 0.51 (the "License"); you may not use this file except in
@@ -10,6 +10,9 @@
 // specific language governing permissions and limitations under the License.
 //
 // Fabian Schuiki <fschuiki@iis.ee.ethz.ch>
+// Florian Zaruba <zarubaf@iis.ee.ethz.ch>
+
+`include "common_cells/registers.svh"
 
 /// A clock domain crossing FIFO, using gray counters.
 ///
@@ -18,17 +21,28 @@
 /// LOG_DEPTH must be at least 1.
 ///
 /// # Constraints
-///
-/// The following constraints need to be set:
-/// - max_delay -from src_wptr_gray_q -to dst_wptr_gray_q
-/// - max_delay -from dst_rptr_gray_q -to src_rptr_gray_q
-/// - max_delay -from fifo_data_q -to fifo_rdata
+/// ```
+/// set_ungroup [get_designs cdc_fifo_gray*] false
+/// set_boundary_optimization [get_designs cdc_fifo_gray*] false
+/// set_max_delay 500 \
+///     -through [get_pins -hierarchical -filter async] \
+///     -through [get_pins -hierarchical -filter async]
+/// set_false_path -hold \
+///     -through [get_pins -hierarchical -filter async] \
+///     -through [get_pins -hierarchical -filter async]
+/// ```
+(* no_ungroup *)
+(* no_boundary_optimization *)
 module cdc_fifo_gray #(
+  /// The width of the default logic type.
+  parameter WIDTH = 1,
   /// The data type of the payload transported by the FIFO.
-  parameter type T = logic,
+  parameter type T = logic [WIDTH-1:0],
   /// The FIFO's depth given as 2**LOG_DEPTH.
-  parameter int LOG_DEPTH = 3
-)(
+  parameter int LOG_DEPTH = 3,
+  /// The number of synchronization registers to insert on the async pointers.
+  parameter int SYNC_STAGES = 2
+) (
   input  logic src_rst_ni,
   input  logic src_clk_i,
   input  T     src_data_i,
@@ -42,117 +56,171 @@ module cdc_fifo_gray #(
   input  logic dst_ready_i
 );
 
+  T [2**LOG_DEPTH-1:0] async_data;
+  logic [LOG_DEPTH:0]  async_wptr;
+  logic [LOG_DEPTH:0]  async_rptr;
+
+  cdc_fifo_gray_src #(
+    .T         ( T         ),
+    .LOG_DEPTH ( LOG_DEPTH )
+  ) i_src (
+    .src_rst_ni,
+    .src_clk_i,
+    .src_data_i,
+    .src_valid_i,
+    .src_ready_o,
+
+    (* async *) .async_data_o ( async_data ),
+    (* async *) .async_wptr_o ( async_wptr ),
+    (* async *) .async_rptr_i ( async_rptr )
+  );
+
+  cdc_fifo_gray_dst #(
+    .T         ( T         ),
+    .LOG_DEPTH ( LOG_DEPTH )
+  ) i_dst (
+    .dst_rst_ni,
+    .dst_clk_i,
+    .dst_data_o,
+    .dst_valid_o,
+    .dst_ready_i,
+
+    (* async *) .async_data_i ( async_data ),
+    (* async *) .async_wptr_i ( async_wptr ),
+    (* async *) .async_rptr_o ( async_rptr )
+  );
+
   // Check the invariants.
-  //pragma translate_off
-  initial begin
-    assert(LOG_DEPTH > 0);
-  end
-  //pragma translate_on
+  // pragma translate_off
+  `ifndef VERILATOR
+  initial assert(LOG_DEPTH > 0);
+  initial assert(SYNC_STAGES >= 2);
+  `endif
+  // pragma translate_on
+
+endmodule
+
+
+(* no_ungroup *)
+(* no_boundary_optimization *)
+module cdc_fifo_gray_src #(
+  parameter type T = logic,
+  parameter int LOG_DEPTH = 3,
+  parameter int SYNC_STAGES = 2
+)(
+  input  logic src_rst_ni,
+  input  logic src_clk_i,
+  input  T     src_data_i,
+  input  logic src_valid_i,
+  output logic src_ready_o,
+
+  output T [2**LOG_DEPTH-1:0] async_data_o,
+  output logic [LOG_DEPTH:0]  async_wptr_o,
+  input  logic [LOG_DEPTH:0]  async_rptr_i
+);
 
   localparam int PTR_WIDTH = LOG_DEPTH+1;
-  typedef logic [PTR_WIDTH-1:0] pointer_t;
-  typedef logic [LOG_DEPTH-1:0] index_t;
+  localparam [PTR_WIDTH-1:0] PTR_FULL = (1 << LOG_DEPTH);
 
-  localparam pointer_t PTR_FULL  = (1 << LOG_DEPTH);
-  localparam pointer_t PTR_EMPTY = '0;
+  T [2**LOG_DEPTH-1:0] data_q;
+  logic [PTR_WIDTH-1:0] wptr_q, wptr_d, wptr_bin, wptr_next, rptr, rptr_bin;
 
-  // Allocate the registers for the FIFO memory with its separate write and read
-  // ports. The FIFO has the following ports:
-  //
-  // - write: fifo_widx, fifo_wdata, fifo_write, src_clk_i
-  // - read: fifo_ridx, fifo_rdata
-  index_t fifo_widx, fifo_ridx;
-  logic fifo_write;
-  T fifo_wdata, fifo_rdata;
-  T fifo_data_q [2**LOG_DEPTH];
-
-  assign fifo_rdata = fifo_data_q[fifo_ridx];
-
-  for (genvar i = 0; i < 2**LOG_DEPTH; i++) begin : g_word
-    always_ff @(posedge src_clk_i, negedge src_rst_ni) begin
-      if (!src_rst_ni)
-        fifo_data_q[i] <= '0;
-      else if (fifo_write && fifo_widx == i)
-        fifo_data_q[i] <= fifo_wdata;
-    end
+  // Data FIFO.
+  assign async_data_o = data_q;
+  for (genvar i = 0; i < 2**LOG_DEPTH; i++) begin : gen_word
+    `FFLNR(data_q[i], src_data_i, src_valid_i & src_ready_o & (wptr_bin[LOG_DEPTH-1:0] == i), src_clk_i)
   end
 
-  // Create the write and read pointers in the source and destination domain.
-  // These are binary counters combined with a Gray encoder. Both the binary and
-  // the Gray coded output are registered; the binary one for use in the local
-  // domain, the Gray one for synchronization into the other domain.
-  pointer_t src_wptr_bin_q, src_wptr_gray_q, dst_rptr_bin_q, dst_rptr_gray_q;
-  pointer_t src_wptr_bin_d, src_wptr_gray_d, dst_rptr_bin_d, dst_rptr_gray_d;
-
-  assign src_wptr_bin_d = src_wptr_bin_q + 1;
-  assign dst_rptr_bin_d = dst_rptr_bin_q + 1;
-
-  binary_to_gray #(PTR_WIDTH) i_src_b2g (src_wptr_bin_d, src_wptr_gray_d);
-  binary_to_gray #(PTR_WIDTH) i_dst_b2g (dst_rptr_bin_d, dst_rptr_gray_d);
-
-  always_ff @(posedge src_clk_i, negedge src_rst_ni) begin
-    if (!src_rst_ni) begin
-      src_wptr_bin_q  <= '0;
-      src_wptr_gray_q <= '0;
-    end else if (src_valid_i && src_ready_o) begin
-      src_wptr_bin_q  <= src_wptr_bin_d;
-      src_wptr_gray_q <= src_wptr_gray_d;
-    end
+  // Read pointer.
+  for (genvar i = 0; i < PTR_WIDTH; i++) begin : gen_sync
+    sync #(.STAGES(SYNC_STAGES)) i_sync (
+      .clk_i    ( src_clk_i       ),
+      .rst_ni   ( src_rst_ni      ),
+      .serial_i ( async_rptr_i[i] ),
+      .serial_o ( rptr[i]         )
+    );
   end
+  gray_to_binary #(PTR_WIDTH) i_rptr_g2b (.A(rptr), .Z(rptr_bin));
 
-  always_ff @(posedge dst_clk_i, negedge dst_rst_ni) begin
-    if (!dst_rst_ni) begin
-      dst_rptr_bin_q  <= '0;
-      dst_rptr_gray_q <= '0;
-    end else if (dst_valid_o && dst_ready_i) begin
-      dst_rptr_bin_q  <= dst_rptr_bin_d;
-      dst_rptr_gray_q <= dst_rptr_gray_d;
-    end
-  end
-
-  // Move the Gray-coded pointers over into the other clock domain and
-  // synchronize them to reduce the probability of metastability.
-  pointer_t src_rptr_gray_q, src_rptr_gray_q2;
-  pointer_t dst_wptr_gray_q, dst_wptr_gray_q2;
-
-  always_ff @(posedge src_clk_i, negedge src_rst_ni) begin
-    if (!src_rst_ni) begin
-      src_rptr_gray_q  <= '0;
-      src_rptr_gray_q2 <= '0;
-    end else begin
-      src_rptr_gray_q  <= dst_rptr_gray_q;
-      src_rptr_gray_q2 <= src_rptr_gray_q;
-    end
-  end
-
-  always_ff @(posedge dst_clk_i, negedge dst_rst_ni) begin
-    if (!dst_rst_ni) begin
-      dst_wptr_gray_q  <= '0;
-      dst_wptr_gray_q2 <= '0;
-    end else begin
-      dst_wptr_gray_q  <= src_wptr_gray_q;
-      dst_wptr_gray_q2 <= dst_wptr_gray_q;
-    end
-  end
-
-  // Reverse the Gray coding of the synchronized pointers.
-  pointer_t src_rptr_bin, dst_wptr_bin;
-
-  gray_to_binary #(PTR_WIDTH) i_src_g2b (src_rptr_gray_q2, src_rptr_bin);
-  gray_to_binary #(PTR_WIDTH) i_dst_g2b (dst_wptr_gray_q2, dst_wptr_bin);
+  // Write pointer.
+  assign wptr_next = wptr_bin+1;
+  gray_to_binary #(PTR_WIDTH) i_wptr_g2b (.A(wptr_q), .Z(wptr_bin));
+  binary_to_gray #(PTR_WIDTH) i_wptr_b2g (.A(wptr_next), .Z(wptr_d));
+  `FFLARN(wptr_q, wptr_d, src_valid_i & src_ready_o, '0, src_clk_i, src_rst_ni)
+  assign async_wptr_o = wptr_q;
 
   // The pointers into the FIFO are one bit wider than the actual address into
   // the FIFO. This makes detecting critical states very simple: if all but the
   // topmost bit of rptr and wptr agree, the FIFO is in a critical state. If the
   // topmost bit is equal, the FIFO is empty, otherwise it is full.
-  assign src_ready_o = ((src_wptr_bin_q ^ src_rptr_bin) != PTR_FULL);
-  assign dst_valid_o = ((dst_rptr_bin_q ^ dst_wptr_bin) != PTR_EMPTY);
+  assign src_ready_o = ((wptr_bin ^ rptr_bin) != PTR_FULL);
 
-  // Drive the FIFO write and read ports.
-  assign fifo_widx  = src_wptr_bin_q;
-  assign fifo_wdata = src_data_i;
-  assign fifo_write = src_valid_i && src_ready_o;
-  assign fifo_ridx  = dst_rptr_bin_q;
-  assign dst_data_o = fifo_rdata;
+endmodule
+
+
+(* no_ungroup *)
+(* no_boundary_optimization *)
+module cdc_fifo_gray_dst #(
+  parameter type T = logic,
+  parameter int LOG_DEPTH = 3,
+  parameter int SYNC_STAGES = 2
+)(
+  input  logic dst_rst_ni,
+  input  logic dst_clk_i,
+  output T     dst_data_o,
+  output logic dst_valid_o,
+  input  logic dst_ready_i,
+
+  input  T [2**LOG_DEPTH-1:0] async_data_i,
+  input  logic [LOG_DEPTH:0]  async_wptr_i,
+  output logic [LOG_DEPTH:0]  async_rptr_o
+);
+
+  localparam int PTR_WIDTH = LOG_DEPTH+1;
+  localparam [PTR_WIDTH-1:0] PTR_EMPTY = '0;
+
+  T dst_data;
+  logic [PTR_WIDTH-1:0] rptr_q, rptr_d, rptr_bin, rptr_bin_d, rptr_next, wptr, wptr_bin;
+  logic dst_valid, dst_ready;
+  // Data selector and register.
+  assign dst_data = async_data_i[rptr_bin[LOG_DEPTH-1:0]];
+
+  // Read pointer.
+  assign rptr_next = rptr_bin+1;
+  gray_to_binary #(PTR_WIDTH) i_rptr_g2b (.A(rptr_q), .Z(rptr_bin));
+  binary_to_gray #(PTR_WIDTH) i_rptr_b2g (.A(rptr_next), .Z(rptr_d));
+  `FFLARN(rptr_q, rptr_d, dst_valid & dst_ready, '0, dst_clk_i, dst_rst_ni)
+  assign async_rptr_o = rptr_q;
+
+  // Write pointer.
+  for (genvar i = 0; i < PTR_WIDTH; i++) begin : gen_sync
+    sync #(.STAGES(SYNC_STAGES)) i_sync (
+      .clk_i    ( dst_clk_i       ),
+      .rst_ni   ( dst_rst_ni      ),
+      .serial_i ( async_wptr_i[i] ),
+      .serial_o ( wptr[i]         )
+    );
+  end
+  gray_to_binary #(PTR_WIDTH) i_wptr_g2b (.A(wptr), .Z(wptr_bin));
+
+  // The pointers into the FIFO are one bit wider than the actual address into
+  // the FIFO. This makes detecting critical states very simple: if all but the
+  // topmost bit of rptr and wptr agree, the FIFO is in a critical state. If the
+  // topmost bit is equal, the FIFO is empty, otherwise it is full.
+  assign dst_valid = ((wptr_bin ^ rptr_bin) != PTR_EMPTY);
+
+  // Cut the combinatorial path with a spill register.
+  spill_register #(
+    .T       ( T           )
+  ) i_spill_register (
+    .clk_i   ( dst_clk_i   ),
+    .rst_ni  ( dst_rst_ni  ),
+    .valid_i ( dst_valid   ),
+    .ready_o ( dst_ready   ),
+    .data_i  ( dst_data    ),
+    .valid_o ( dst_valid_o ),
+    .ready_i ( dst_ready_i ),
+    .data_o  ( dst_data_o  )
+  );
 
 endmodule